+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 ### Updated
-
+- Updated newrelic implementation
 - Refactored collection/item lanes into one `Lane` component (DR-3191)
 - Updated header on /divisions/:slug page to have a tabIndex of -1 (DR-3229)